uop: move pm_mops to movement, breaking codegen-rangeify cycle
